+++ Frends.AzureDataLake.DownloadFiles/Frends.AzureDataLake.DownloadFiles.Tests/tests/DownloadingTests.cs
@@ -18,8 +18,11 @@ public class DownloadingTests : TestsBase
     public async Task ThrowIfDestinationDirectoryDoesNotExist()
     {
         await AzureDataLake.DownloadFiles(
-            new Source { ConnectionString = connectionString, ContainerName = containerName },
-            new Destination { Directory = "C:/NonExistingDir/ForSure/test" },
+            new Input
+            {
+                Source = new Source { ConnectionString = connectionString, ContainerName = containerName },
+                Destination = new Destination { Directory = "C:/NonExistingDir/ForSure/test" }
+            },
             new Options(),
             CancellationToken.None
         );
@@ -29,17 +32,20 @@ public class DownloadingTests : TestsBase
     public async Task DownloadSingleFileWithConnectionString()
     {
         var result = await AzureDataLake.DownloadFiles(
-            new Source
+            new Input
             {
-                ConnectionString = connectionString,
-                ContainerName = containerName,
-                FilePattern = file1a
+                Source = new Source
+                {
+                    ConnectionString = connectionString,
+                    ContainerName = containerName,
+                    FilePattern = file1a
+                },
+                Destination = new Destination { Directory = testDirectory }
             },
-            new Destination { Directory = testDirectory },
             new Options(),
             CancellationToken.None
         );
-        Assert.IsTrue(result.IsSuccess);
+        Assert.IsTrue(result.Success);
         CollectionAssert.AreEqual(result.DownladedFiles, SingleFileResult.DownladedFiles);
         Assert.That.FileExists(Path.Combine(testDirectory, file1a));
     }
@@ -48,21 +54,24 @@ public class DownloadingTests : TestsBase
     public async Task DownloadSingleFileWithOAuth2()
     {
         var result = await AzureDataLake.DownloadFiles(
-            new Source
+            new Input
             {
-                ConnectionMethod = ConnectionMethod.OAuth2,
-                ContainerName = containerName,
-                StorageAccountName = storageAccount,
-                ApplicationID = appID,
-                TenantID = tenantID,
-                ClientSecret = clientSecret,
-                FilePattern = file1a
+                Source = new Source
+                {
+                    ConnectionMethod = ConnectionMethod.OAuth2,
+                    ContainerName = containerName,
+                    StorageAccountName = storageAccount,
+                    ApplicationID = appID,
+                    TenantID = tenantID,
+                    ClientSecret = clientSecret,
+                    FilePattern = file1a
+                },
+                Destination = new Destination { Directory = testDirectory }
             },
-            new Destination { Directory = testDirectory },
             new Options(),
             CancellationToken.None
         );
-        Assert.IsTrue(result.IsSuccess);
+        Assert.IsTrue(result.Success);
         CollectionAssert.AreEqual(result.DownladedFiles, SingleFileResult.DownladedFiles);
         Assert.That.FileExists(Path.Combine(testDirectory, file1a));
     }
@@ -71,17 +80,20 @@ public class DownloadingTests : TestsBase
     public async Task DownloadNothingIfPatternFileNotFound()
     {
         var result = await AzureDataLake.DownloadFiles(
-            new Source
+            new Input
             {
-                ConnectionString = connectionString,
-                ContainerName = containerName,
-                FilePattern = "nonExisting.txt"
+                Source = new Source
+                {
+                    ConnectionString = connectionString,
+                    ContainerName = containerName,
+                    FilePattern = "nonExisting.txt"
+                },
+                Destination = new Destination { Directory = testDirectory }
             },
-            new Destination { Directory = testDirectory },
             new Options(),
             CancellationToken.None
         );
-        Assert.IsTrue(result.IsSuccess);
+        Assert.IsTrue(result.Success);
         CollectionAssert.AreEqual(result.DownladedFiles, new Dictionary<string, string>());
     }
 
@@ -89,17 +101,20 @@ public class DownloadingTests : TestsBase
     public async Task DownloadMultipleFiles()
     {
         var result = await AzureDataLake.DownloadFiles(
-            new Source
+            new Input
             {
-                ConnectionString = connectionString,
-                ContainerName = containerName,
-                FilePattern = multiFilePatten
+                Source = new Source
+                {
+                    ConnectionString = connectionString,
+                    ContainerName = containerName,
+                    FilePattern = multiFilePatten
+                },
+                Destination = new Destination { Directory = testDirectory }
             },
-            new Destination { Directory = testDirectory },
             new Options(),
             CancellationToken.None
         );
-        Assert.IsTrue(result.IsSuccess);
+        Assert.IsTrue(result.Success);
 
         var orderedResultFiles = result
             .DownladedFiles.OrderBy(x => x.Key)
@@ -127,18 +142,21 @@ public class DownloadingTests : TestsBase
 
         //act
         var result = await AzureDataLake.DownloadFiles(
-            new Source
+            new Input
             {
-                ConnectionString = connectionString,
-                ContainerName = containerName,
-                FilePattern = file1a
+                Source = new Source
+                {
+                    ConnectionString = connectionString,
+                    ContainerName = containerName,
+                    FilePattern = file1a
+                },
+                Destination = new Destination { Directory = testDirectory, Overwrite = true }
             },
-            new Destination { Directory = testDirectory, Overwrite = true },
             new Options(),
             CancellationToken.None
         );
         //assert
-        Assert.IsTrue(result.IsSuccess);
+        Assert.IsTrue(result.Success);
         CollectionAssert.AreEqual(result.DownladedFiles, SingleFileResult.DownladedFiles);
         Assert.That.FileContainsText(Path.Combine(testDirectory, file1a), string.Empty);
     }
@@ -154,17 +172,20 @@ public class DownloadingTests : TestsBase
         }
 
         var result = await AzureDataLake.DownloadFiles(
-            new Source
+            new Input
             {
-                ConnectionString = connectionString,
-                ContainerName = containerName,
-                FilePattern = file2
+                Source = new Source
+                {
+                    ConnectionString = connectionString,
+                    ContainerName = containerName,
+                    FilePattern = file2
+                },
+                Destination = new Destination { Directory = testDirectory, Overwrite = false }
             },
-            new Destination { Directory = testDirectory, Overwrite = false },
             new Options(),
             CancellationToken.None
         );
-        Assert.IsTrue(result.IsSuccess);
+        Assert.IsTrue(result.Success);
         CollectionAssert.AreEqual(result.DownladedFiles, FileAlreadyExistsResult.DownladedFiles);
         Assert.That.FileContainsText(Path.Combine(testDirectory, file2), startContent);
     }