+++ Frends.AzureDataLake.DownloadFiles/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.0] - 2026-08-13
+### Changed [Breaking Change]
+- Task now accepts a single `Input` parameter containing `Source` and `Destination` instead of separate parameters, to comply with Frends task standards.
+- Result property `IsSuccess` has been renamed to `Success`.
+- Result property `ErrorMessage` has been removed. Error details are now returned in the `Error` object (with `Message` and `AdditionalInfo` properties) when `ThrowErrorOnFailure` is false.
+- Added `ErrorMessageOnFailure` option: you can now provide a custom error message that will be used when the task fails.
+- Upgraded target framework from net6.0 to net8.0.
+
 ## [1.2.0] - 2026-01-15
 ### Changed
 - Updated Azure packages to latest versions: